move pathstep, create path
