+++ xrpl/models/__init__.py
@@ -1,5 +1,14 @@
 """Top-level exports for the models package."""
 from xrpl.models import amounts, currencies, requests, transactions
 from xrpl.models.exceptions import XRPLModelException
+from xrpl.models.path import Path, PathStep
 
-__all__ = ["XRPLModelException", "amounts", "currencies", "requests", "transactions"]
+__all__ = [
+    "XRPLModelException",
+    "amounts",
+    "currencies",
+    "requests",
+    "transactions",
+    "Path",
+    "PathStep",
+]
