+++ xrpl/models/transactions/payment.py
@@ -6,7 +6,7 @@ from enum import Enum
 from typing import Dict, List, Optional
 
 from xrpl.models.amounts import Amount, is_xrp
-from xrpl.models.requests import PathStep
+from xrpl.models.path import Path
 from xrpl.models.required import REQUIRED
 from xrpl.models.transactions.transaction import Transaction
 from xrpl.models.transactions.types import TransactionType
@@ -68,7 +68,7 @@ class Payment(Transaction):
     this Check.
     """
 
-    paths: Optional[List[List[PathStep]]] = None
+    paths: Optional[List[Path]] = None
     """
     Array of payment paths to be used (for a cross-currency payment). Must be
     omitted for XRP-to-XRP transactions.
